added bug fixes and comments for download_gedi

+++ access_gedi/download_gedi.py
@@ -10,6 +10,7 @@ maap = MAAP(maap_host="api.maap-project.org")
 
 
 def gedi_filename_to_s3_url(filename: str) -> str:
+    """Convert a GEDI filename to an s3 URL"""
     name_components = filename.split("_")
 
     if name_components[0:2] == ["GEDI01", "B"]:
@@ -33,27 +34,12 @@ def gedi_filename_to_s3_url(filename: str) -> str:
 
     return f"{base_s3}/{basename}/{filename}"
 
-
-def copy_gedi_file(s3_url: str,
-                   s3: fsspec.filesystem) -> str:
-    basename = os.path.basename(s3_url)
-    output_path = f"/projects/my-private-bucket/{basename}"
-    print("Copying file...")
-    s3.get(s3_url, output_path)
-    print("File copied.")
-    return output_path
-
-def get_gedi_data(filename: str, retries: int = 3):
-    attempt = 0
-    delay = 5  
-
-    # Define credentials inside the function
+def open_s3_session():
+    """Get a new session token from MAAP and open an s3 filesystem"""
     credentials = maap.aws.earthdata_s3_credentials(
         "https://data.lpdaac.earthdatacloud.nasa.gov/s3credentials"
     )
-
-    s3_url = gedi_filename_to_s3_url(filename)
-
+    
     s3 = fsspec.filesystem(
         "s3",
         key=credentials['accessKeyId'],
@@ -61,14 +47,35 @@ def get_gedi_data(filename: str, retries: int = 3):
         token=credentials['sessionToken']
     )
 
-    output_path = f"/projects/my-private-bucket/{filename}"
+    return s3
 
+
+def get_gedi_data(filename: str,
+                  target_dir: str,
+                  retries: int = 3):
+    """Download a GEDI file from the MAAP s3 bucket"""
+
+    attempt = 0
+    delay = 5
+
+    # Open the s3 filesystem and get the s3 URL
+    s3 = open_s3_session()
+    s3_url = gedi_filename_to_s3_url(filename)
+
+    # Prepare the output path
+    if not os.path.exists(target_dir):
+        os.makedirs(target_dir)
+    output_path = os.path.join(target_dir, filename)
+    if os.path.exists(output_path):
+        raise FileExistsError(f"File already exists at {output_path}")
+
+    # Download the file
     while attempt < retries:
         try:
+            # Use a temporary file to avoid partial downloads
             with tempfile.NamedTemporaryFile() as temp_fp:
                 s3.get(s3_url, temp_fp.name)
                 shutil.move(temp_fp.name, output_path)
-                break
         except Exception as e:
             attempt += 1
             print(f"Attempt {attempt} failed: {e}")
